Back/requirements.txt
fastapi
uvicorn
langchain_openai
langchain_community
langchain_huggingface
sentence-transformers